Make height float

+++ db/migrate/20231216004215_devise_create_users.rb
@@ -32,7 +32,7 @@ class DeviseCreateUsers < ActiveRecord::Migration[7.1]
       # t.string   :unlock_token # Only if unlock strategy is :email or :both
       # t.datetime :locked_at
 
-      t.integer :height
+      t.float :height
       t.integer :carb
       t.integer :fat
       t.integer :protein
